+++ src/App.js
@@ -1,7 +1,8 @@
 import React from "react";
 import Topbar from "./components/topbar/Topbar";
 import MainSideBar from "./components/mainSideBar/MainSideBar";
-import HomeDashBoard from "./components/homeDashBoard/HomeDashBoard";
+// import HomeDashBoard from "./components/homeDashBoard/HomeDashBoard";
+import HomeDashBoard from "./pages/homeDashBoard/HomeDashBoard"
 import "./App.css"
